Refactor GetSendBuf implementation
This commit refactors GetSendBuf implementation to not use any
heuristics to decide when to return; instead, it will simply wait
until data is available in the channel, and exhausts the channel
before returning from the call. This removes the need to pass a "size
hint" as a parameter.

This change was made after making a few observations:

 - The implementation of tls.Conn.Write itself does not return until
   all of the data has been read in and processed.

 - Past the initial handshake establishment in securesession.go, we
   only execute in a single thread. This fact combined with the
   previous observation suggests that we should be able to swap over
   to this newer implementation with any hassle.

 - The GetSendBuf calls that *are* part of the handshake establishment
   contain messages that are sufficiently small that they fit in a
   single write to the TLS implementation. In addition, nothing on the
   caller side attempts to write to the TLS buffer simultaneously, so
   there shouldn't be any synchronization issues.

Finally, the GetSendBuf methods were renamed to DrainSendBuf to more
accurately reflect their behaviour.

PiperOrigin-RevId: 392720948
Change-Id: I91a41e40e48a786ca5add8bbd91db6b3185fc0aa