Support wrapping shares with asymmetric keys
This commit adds support to STET for encrypting key shares using a
specified public RSA key, rather than specifying a URI to a key stored
in an EKM. This is done by instead specifying a fingerprint of the
public key as the KekInfo, and adding a stanza to the StetConfig that
defines paths to public and private key files.

Some additional refactoring was also done to avoid initializing the
KMS client on all wrap/unwrap operations, since with the addition of
the ability to do offline share wrapping/unwrapping, contacting Cloud
KMS will not always be necessary.

Tested:
 - Generated a private/public RSA keypair via the following:

     $ openssl genrsa -out stet.pem 2048
     $ openssl rsa -in stet.pem -pubout -out stet.crt

   And a fingerprint using the following:

     $ openssl rsa -in stet.pem -pubout -outform DER | \
         openssl sha256 -binary | openssl base64

   Then, copied the fingerprint and paths into a STET config file
   similar to the one defined in examples/offline.yaml, and
   invoked STET, successfully encrypting and decrypting.

 - Did the above swapping out the private key for a newly generated
   one, and the decryption fails as expected.

 - Copied source over to a Confidential VM and built, and all of the
   same tests succeeded and failed as expected.

 - Go toolchain-based tests pass as well as Bazel tests.

 - Existing config files still work, despite the kek_uri field now
   turning into a oneof.

 - TestWrapUnwrapShareAsymmetricKey[Error] tests wrapping/unwrapping.

NOKEYCHECK

PiperOrigin-RevId: 393201239
Change-Id: Ib14ab830b1b6e30cb51c468284f216d4855a02d5