De-escalate root privileges until needed
This commit updates STET to better handle user privileges when invoked
as root. Running the entirety of the binary as root puts us at risk of
more security vulnerabilities, and causes the more pressing issue of
writing output files as root rather than the user who invoked it.

With this approach, the STET binary immediately checks if its effective
UID is root, and if so, swaps it with the real UID to run as though it
didn't have these privileges. Only when the TPM device needs to be
accessed do we reverse this swap to run as root.

Tested:

Built STET and changed file permissions by performing the following:

 $ sudo chown root /usr/local/bin/stet
 $ sudo chmod u+sx,a+rx /usr/local/bin/stet

Then, performed an encryption via `stet encrypt ...`; because this was
running on a Confidential VM, the vTPM device was successfully able to
produce an attestation, and the resulting output file was owned by me.

In addition, commenting out the calls to swapEffectiveAndRealUID()
that occur directly before attempting to call tpm2.OpenTPM() result
in that call failing, indicating that we are indeed escalating and
de-escalating root privileges correctly.

Finally, executing STET normally in a non-CVM environment where the
binary is just invoked with user privileges works as expected.

PiperOrigin-RevId: 393997863
Change-Id: I4b762b3900ed268f7ccbcba3a5b52a3d3babd1ac