Update evidence session binding logic
This commit updates the Finalize step of the session establishment flow
to align closer to clarifications that were made to the protocol spec.

In particular, the purpose string "TLSAttestationV1" was previously
prepended as bytes to the attestation produced to the vTPM, before
being session-encrypted. However, this should instead be included
in the nonce used to generate the attestation.

Note that in cases where no attestation is being provided by the
client, the `attestation_evidence_records` field is now left empty;
currently, some session-encrypted records were being sent as part
of the request in this situation.

This change is justified because, without producing an attestation,
protection against a MitM is not afforded by simply transmitting the
keying material to the server. Therefore, it's simpler to just leave
the field empty.

To ensure tests still pass, the reference server implementation was
also updated to reflect the change to the null attestation flow (ie.
not session-decrypting an unmarshaling an empty set of records).

PiperOrigin-RevId: 395791986
Change-Id: Ifb7410429008d27599939840c03488b9ff13e954