Display external key URI used for unwrapping
This commit updates STET to display the underlying key URI when an
external key is used, rather than the URI that is specified in the
configuration file (which is a Cloud KMS key).

This is an important change within the framework of auditability;
this allows users to verify that Cloud KMS pointed them to the EKM
that they expected to contact, rather than misdirecting them.

PiperOrigin-RevId: 396861318
Change-Id: I72994ab6290ce6cb8b91e8add3fa197d117af6c4