Add version subcommand for debugging
This commit adds a new version subcommand to STET, making it easier
for users to determine what build of STET they are currently running.
This was implemented as a subcommand rather than a --version flag due
to how the subcommands library works (it's simpler to just add a new
subcommand, rather than a flag that works globally across commands).

PiperOrigin-RevId: 396922620
Change-Id: I05a04efb84099df7acdd3c9498b150106b85a4c6