Return error if request contains no TLS records
This commit updates the non-Finalize endpoints in the secure session
server implementation to return an error if no records were passed
with the RPC request. Without this, an empty byte slice is passed to
shim.QueueReceiveBuf(), and the subsequent call to shim.DrainSendBuf()
hangs due to the TLS implementation not doing anything.

This change isn't needed in the Finalize implementation as an empty
records field indicates a null attestation, which is intentional.

PiperOrigin-RevId: 397419807
Change-Id: I1a3e14485f111e968967c0c71ff4074a1133bc16