Fix decryption flow when using HTTP client
This commit fixes a bug with establishing the secure session while
unwrapping a key. Rather than passing in `addr` (which is just the
hostname of the key URI), the entire key URI should be passed in,
otherwise the logic in confidentialekmclient.go is unable to properly
form the endpoints required to establish the secure session.

PiperOrigin-RevId: 397427188
Change-Id: Icc50b3ddb0d1b72377c949e9b168333ddbf2b915