Use readers/writers for AEAD operations
This commit updates the AEAD portion of the codebase to leverage the
ability to pass an io.Reader / io.Writer to Tink's cipher methods. By
updating the client.Encrypt and client.Decrypt methods to take in
readers and writers, we can have Tink read and write data directly to
disk or stdin/stdout as it's ready to do so.

Note that after this change is made, the ciphertext is still held
in memory (as it is serialized to a proto with the metadata before
being written to disk), but it avoids having 2x the size of the
ciphertext in memory. A future commit will update the encryption
format to avoid having to do this by being able to seek directly
to the ciphertext portion of the encrypted data blob.

Tested:
 - Unit tests pass.
 - Ran `stet encrypt` and `stet decrypt` end-to-end successfully.
PiperOrigin-RevId: 397854453
Change-Id: Ifde41fec75a0cb54fecce1a6b819636c8dd76860