Create STET encrypted file format
This commit introduces a new method of storing data encrypted by STET
on disk. Rather than a serialized proto with one field for ciphertext
and one for metadata, this new approach contains a header, metadata
proto, and ciphertext concatenated to one another in that order.

The benefit of this approach is that, on the encryption side, a single
io.Writer can be used to write the header followed by the metadata,
then the open io.Writer is passed to Tink to AEAD-encrypt data
directly to disk or stdout, without needing an intermediary buffer to
read the entire payload into memory. The inverse is done with an
io.Reader on the decryption side of things, first reading the header
and metadata explicitly, then passing the "rest of the file" to Tink
for ciphertext decryption.

Tested:
 - Unit tests pass.
 - End-to-end test of `stet encrypt` and `stet decrypt` succeeds.
 - Used xxd to verify header format:

                   magic string         len(metadata)
          |------------------------------|   |--|
00000000: 5354 4554 454e 4352 5950 5445 4401 ee00  STETENCRYPTED...
                                          ||
                                    format version
PiperOrigin-RevId: 398079888
Change-Id: I0282bbf2bf69c89c972637931082bd7838f1eed0