Rename STET proto packages
This commit updates the package names of all protos to align with the
protobuf style guide's recommendations [1]. This required renaming
the "UNKNOWN" enum options in order to avoid symbol conflicts.

The old names for confidential_wrap.proto and secure_session.proto was
a holdover from the initial version of STET that communicated with
EKMs over gRPC, rather than HTTP. The package name of the RPC service
definitions had to match on both the client and server side of things,
and therefore we were unable to pick our own package name whilst
maintaining compatibility.

Now that we communicate over HTTP requests, and the proto messages
are just serialized and used as the request body, it doesn't matter if
there is a package name mismatch between the client and server.

[1] https://developers.google.com/protocol-buffers/docs/style#packages

PiperOrigin-RevId: 398363047
Change-Id: I675b8217c1e04630adff50c23d8d550de61bb987