Add checks for nil EncryptConfig/DecryptConfig
This commit updates the Encrypt() and Decrypt() methods to check
for a nil EncryptConfig or DecryptConfig respectively. This prevents
a potential nil pointer dereference from occurring if the STET binary
calls invokes those methods, but the config file doesn't have the
corresponding EncryptConfig or DecryptConfig stanzas.

This check is also explicitly added to the STET binary to produce
more informative error messages on the missing config stanzas.

PiperOrigin-RevId: 398378701
Change-Id: Ic957db2ef3702a180eb7f2ca1ad6294162b52396