Change key URIs to key paths in reference server
This commit changes the reference server's key URIs into key paths,
since the relevant WrapRequest field is a key path, not a key URI.

Also, this commit updates the default key path and resource name flags
for the secure session binary to be one of the default key paths that
are hard-coded into the reference server implementation.

This allows both the handshake and the wrapping/unwrapping flows to
succeed when using the securesession binary as a conformance check
against the reference server binary, without needing to guess at what
key paths are valid for the reference server.

PiperOrigin-RevId: 398598543
Change-Id: I5095bba8d932f7d97704bb202a8d627a84545ee0