Lengthen exported keying material to 32 bytes
This commit lengthens the exported keying material used in the nonce
when generating attestations from 16 bytes to 32 bytes. This was done
in order to mitigate the risk of an attacker opening TLS connections
in the hope that they hit the same random exported key.

By moving from 16 to 32 bytes of keying material, an attacker would
need to try somewhere on the order of 2^128 connections (virtually
impossible), rather than on the order of 2^64 (extremely difficult).

PiperOrigin-RevId: 399777478
Change-Id: Ibbf368863a1be3594a40a2b76dd049509a3ee039