Add restricted TLS 1.2 support
This commit adds support for using TLS 1.2 in the secure session to
both the STET client and server, with a restriction on the allowed
cipher suites. The cipher suites chosen are exactly those that are
currently allowable in TLS 1.3.

This change is being made to accommodate EKM partners that may not be
able to use TLS 1.3 due to technical limitations.

To assist with testing this change, a `--tls12` flag was added to the
server binary to run the reference server in a mode where its inner
TLS session only supports TLS 1.2, rather than TLS 1.3.

Because the set of allowable cipher suites is restricted to elliptic
curve algorithms rather than RSA, the test key and certificate used
in the reference server were regenerated using an EC private key.

As a side effect of verifying this change, the explicit call to
client.handshake() in securesession.go that occurred outside of the
repeated handshake loop was removed, as it was confirmed unnecessary.

In addition, the last usage of the non-blocking DrainSendBuf() method
was deemed unnecessary after closer investigation of the divergence
between TLS 1.2 and 1.3 in the Handshake request, so it was removed
from the transport shim altogether.

Tested:
 - Unit tests pass.
 - Confirmed `stet encrypt` still establishes secure session and can
   encrypt data to an EKM partner with existing TLS 1.3-based server.
 - The securesession binary was used to verify the handshake against
   the server binary locally, in both TLS 1.2 and 1.3 modes.

NOKEYCHECK

PiperOrigin-RevId: 399784375
Change-Id: Ib18fa9902a2c7025870bda9671ce6b5d133a6429