Implement nonce type negotiation for client
This commit implements logic for negotiating the "nonce type" used by
the client to generate attestations. This is done by taking advantage
of the already-existing negotiate attestation payload send as part of
the secure session establishment.

This is implemented in a way that allows clients and servers to be
updated independently of one another; a single client can attest in
different ways to different servers as part of a single invocation,
ensuring that separate KMS providers do not have to coordinate
updating to the newest version of the nonce type, nor drop support
for legacy versions at the same time.

In addition, the client logic is done in such a way that if the
server responds with an empty list of nonce types, it is treated as
though the server responded with the basic "32-byte exported keying
material nonce" that was originally codified into the protocol. This
situation applies if any servers have not been updated to perform
nonce type negotiation.

PiperOrigin-RevId: 401314860
Change-Id: I4cb56fff9487fd9ec18eaa3decfc69f441a905d5