Fix logic for extracting unwrapped share URIs
The existing logic was a holdover from the URI was only inferred from
the KekInfo. Now the unwrapping methods return the URI used (in the
case of a key stored in an external key manager, the actual external
key URI rather than the URI stored in the configuration).

This refactoring also gets rid of a bug that could occur if unwrapping
failed for certain shares, causing the indexing of unwrappedShares to
not align with that of the KekInfos.

PiperOrigin-RevId: 401586002
Change-Id: I972b79e0c858de71014c67ef4d7cec2f2f54ed49