Update attestation negotiation logic
This commit updates the reference server implementation to not treat
the null attestation evidence type differently from other evidence
types. If the client specifies that it supports the null attestation
(which in virtually all real-world situations, it should), then the
server should respond appropriately.

It also correspondingly updates the secure session implementation to
explicitly provide the null attestation type, even when running on a
machine capable of producing vTPM attestations.

Finally, it fixes a bug where the switch statement in the server
implementation was written in a way that didn't cause the negotiated
type to be appended when matching an expected type (in Go, multiple
cases for the same block of execution must be comma-separated).

PiperOrigin-RevId: 421145881
Change-Id: I5b56ae3122606f8c545582a17011e2ec9fd6f762