Add NegotiateAttestation conformance tests
This commit adds conformance tests for negotiating attestation types.
Specifically, the AttestationEvidenceTypeList that gets sent with the
client's request contains a list of attestation evidence types and
nonce types that the client knows of, and the test suite ensures the
server responds appropriately (with some subset of types that *it*
can handle).

Note that nonce negotiation is handled in a special manner, where the
conformance suite does not enforce that nonce types are negotiated
correctly. This is because the part of the protocol that describes
negotiating nonce types has not been formalized for a long enough
period of time to expect all server implementations to correctly
handle this at this point.

PiperOrigin-RevId: 421149798
Change-Id: Iff79d5fa0de850945d03c63e811f111d976a4ba8