Return error on failed attestation negotiation
This commit updates the reference server implementation to return an
error if the set of attestation evidence types provided in the client
request is completely disjoint from the set of evidence types that
the server supports.

Attempting to return a 0-length list in this situation causes issues
when the marshaled proto is passed to the TLS implementation; the
server hangs because it waits for data to be written back to the
TLS intercept, but the marshaled proto is 0 bytes long, so nothing
ever gets written.

PiperOrigin-RevId: 421355632
Change-Id: Ia5a5f7402181f58c5baf4146ebbdf224cb5e6a89