Add Finalize tests to conformance suite
This commit adds tests for the Finalize endpoint to the conformance
test suite. It leverages the `client.EstablishSecureSession()` method
for testing the happy path when generating vTPM attestations, in order
to avoid duplicating the entire implementation.

In addition, this commit cleans up other parts of the test suite to
not have to require specifying `identityFn` on every test case when
not mutating the underlying TLS records or session context.

PiperOrigin-RevId: 421933946
Change-Id: Ie58b0408a4f1b8ecd468f240738ba736e829f3d9