Error if TPM 2 Quote and TCG Event Log not linked
Has the client error if, while negotiating attestation, the server
only requests the TPM 2 Quote or only requests the TCG Event Log. The
implementation currently only supports sending both, this change just
makes that more explicit by erroring if the server only requests one.
For a meaningful vTPM attestation, we recommended that both should be
present, since direct use the values of the PCR registers is brittle
and susceptible to race conditions in measurement ordering.

Adds corresponding changes to the reference server and the conformance
test.

PiperOrigin-RevId: 431480763
Change-Id: Ib0c67cd4a5151f31415b0e3742297119cd5fe37e