Add JWT tests and fix bugs in conformance tests
From this commit on, in order to run the conformance tests a valid
Google access token will need to be referenced by the file at the
GOOGLE_APPLICATION_CREDENTIALS environment variable.

This change also updates the reference server to verify the JWT, based
on an audience field that is passed in at construction.

PiperOrigin-RevId: 553914967
Change-Id: I482ff8a63a4d2c836a24d0b3cbc369ab78d8bef0